refactored imaging astar module and added separate imaging module
